Refactor i686 architecture initialization and cleanup magic numbers

+++ src/arch/i686/pit.c
@@ -19,8 +19,8 @@
  */
 void tmr_init(void) {
     outb(PIT_MODE_CMD, PIT_CHAN0 | PIT_LOHIGH | PIT_MODE2);
-    outb(PIT_CHAN0_DATA, (uint8_t)(DIVISOR & 0xff));
-    outb(PIT_CHAN0_DATA, (uint8_t)((DIVISOR & 0xff00) >> 8));
+    outb(PIT_CHAN0_DATA, DIVISOR & 0xff);
+    outb(PIT_CHAN0_DATA, (DIVISOR & 0xff00) >> 8);
 }
 
 extern struct k_ctx ctx;
@@ -31,8 +31,8 @@ extern struct k_ctx ctx;
  * @return Returns 1 if a context switch should occur, otherwise 0.
  */
 bool tmr_pit_handler(void) {
-    outb(PIT_CHAN0_DATA, (uint8_t)(DIVISOR & 0xff));
-    outb(PIT_CHAN0_DATA, (uint8_t)((DIVISOR & 0xff00) >> 8));
+    outb(PIT_CHAN0_DATA, DIVISOR & 0xff);
+    outb(PIT_CHAN0_DATA, (DIVISOR & 0xff00) >> 8);
     pic_send_eoi(0);
 
     ctx.ticks++;
